pyproject.toml
[project]
name = "momentum-trading"
version = "0.1.0"
description = "Momentum Trading - Hyperliquid"
authors = [
    {name = "Gajesh Naik",email = "26431906+Gajesh2007@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "hyperliquid-python-sdk (>=0.19.0,<0.20.0)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "eth-account (>=0.13.7,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
